requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.7.2
certifi>=2021.5.30
charset-normalizer>=2.0.2
idna>=3.2
requests>=2.26.0
urllib3>=1.26.6

